--- common.go
@@ -1,28 +1,0 @@
-package stream
-
-// Metadata for a release or stream
-type Metadata struct {
-	LastModified string `json:"last-modified"`
-}
-
-// ImageFormat contains all artifacts for a single OS image
-type ImageFormat struct {
-	Disk      *Artifact `json:"disk,omitempty"`
-	Kernel    *Artifact `json:"kernel,omitempty"`
-	Initramfs *Artifact `json:"initramfs,omitempty"`
-	Rootfs    *Artifact `json:"rootfs,omitempty"`
-}
-
-// Artifact represents one image file, plus its metadata
-type Artifact struct {
-	Location  string `json:"location"`
-	Signature string `json:"signature"`
-	Sha256    string `json:"sha256"`
-}
-
-// GcpImage represents a GCP cloud image
-type GcpImage struct {
-	Project string `json:"project,omitempty"`
-	Family  string `json:"family,omitempty"`
-	Name    string `json:"name,omitempty"`
-}
