+++ release/release_test.go
@@ -1,4 +1,4 @@
-package stream
+package release
 
 import (
 	"encoding/json"
