+++ stream/fixtures/fcos-stream.json
